requirements.txt
annotated-types==0.7.0
anyio==4.11.0
astroid==4.0.1
certifi==2025.10.5
coverage==7.11.0
dill==0.4.0
distro==1.9.0
fastapi==0.115.6
uvicorn[standard]==0.34.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
isort==7.0.0
jiter==0.11.0
mccabe==0.7.0
openai==2.6.0
packaging==25.0
platformdirs==4.5.0
pluggy==1.6.0
pydantic==2.12.0
pydantic_core==2.41.1
Pygments==2.19.2
pylint==4.0.2
pytest==8.4.2
pytest-asyncio==0.25.0
pytest-cov==7.0.0
pytest-mock==3.15.1
pytest-pylint==0.21.0
python-dotenv==1.1.1
sniffio==1.3.1
tomlkit==0.13.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0

# STUDENT ASSISTANT FEATURE - Document processing dependencies
aiofiles==24.1.0
python-multipart==0.0.9
PyPDF2==3.0.1
python-docx==1.1.2
ebooklib==0.18
chromadb==0.5.23
sentence-transformers==3.1.1
pydub==0.25.1
